appmenus: use Admin API (part 2)
Change the appmenus code to work independent of qubesd process, in user
session. This is preparation for GUI VM. For now, all operations are
still triggered by core extension, but the code is structured to
trivially change it to qrexec service or such, later.
Also store data in user home, independent of /var/lib/qubes (which
persist only in dom0).